+++ src/main/cpp/operators/support/secure_sort_condition.cpp
@@ -11,7 +11,7 @@ void SecureSortCondition::compareAndSwap(QueryTuple &lhs, QueryTuple &rhs) {
     emp::Bit falseBit(false);
 
 
-    for(int i = 0; i < sortDefinition.size(); ++i) {
+    for(size_t i = 0; i < sortDefinition.size(); ++i) {
         types::Value lhsValue = SortCondition::getValue(lhs, sortDefinition[i]);
         types::Value rhsValue = SortCondition::getValue(rhs, sortDefinition[i]);
